refactor(classifiers): Rename scales to classifiers to avoid confusion, and add comments.
