+++ lib/classifiers/__tests__/diverging.js
@@ -3,7 +3,7 @@
 import diverging from '../diverging'
 import equal from '../equal-interval'
 
-describe('scales > diverging', () => {
+describe('classifiers > diverging', () => {
   it('should work', () => {
     const data = new Int32Array(900)
     // yields even dispersion between -4 and +4
