+++ lib/classifiers/__tests__/equal-interval.js
@@ -4,7 +4,7 @@ import {mockGrid} from '../../mock-data'
 import equalInterval from '../equal-interval'
 import {scaleLog} from 'd3-scale'
 
-describe('scales > equal-interval', () => {
+describe('classifiers > equal-interval', () => {
   it('should handle a linear classifier', () => {
     expect(equalInterval()(mockGrid, 5)).toMatchSnapshot()
   })
