+++ lib/classifiers/__tests__/quantile.js
@@ -2,7 +2,7 @@
 
 import quantile from '../quantile'
 
-describe('scales > quantile', () => {
+describe('classifiers > quantile', () => {
   it('should work', () => {
     const data = new Int32Array(1000)
     // spread into 10 classes evenly
