+++ lib/classifiers/diverging.js
@@ -4,11 +4,11 @@ import equal from './equal-interval'
 import range from 'lodash.range'
 
 /**
- * Create a diverging scale from any other scale, by setting the scheme to another scale.
+ * Create a diverging classifier from any other classifier,
+ * set scheme to the other classifier.
  * If the center is not zero, specify the center parameter.
- * If the scale you are using requires other options, pass them in as well.
+ * If the classifier you are using requires other options, pass them in as well.
  */
-
 const diverging = ({ scheme = equal, center = 0, ...opts } = {}) => (grid, nBreaks) => {
   if (nBreaks % 2 === 0) nBreaks-- // we need an odd number of clases
