+++ lib/classifiers/equal-interval.js
@@ -1,7 +1,11 @@
 import {scaleLinear} from 'd3-scale'
 import range from 'lodash.range'
 
-/** equal interval classifier, potentially in a space other than the linear */
+/**
+ * equal interval classifier, potentially in a space other than the linear
+ * Passing in no arguments will use a linear space. you can also pass in
+ * a { scale } parameter, which should be a d3 continuous scale with the domain set correctly.
+ */
 const equal = ({ scale } = {}) => (grid, nBreaks) => {
   // copy reference so that it does not persist between calls to thunk
   let localScale = scale
