+++ lib/classifiers/quantile.js
@@ -1,5 +1,6 @@
 import XorShift from 'xorshift'
 
+/** A quantile classifier. Pass in an optional no data value */
 const quantile = ({ noDataValue = null }) => (grid, nBreaks) => {
   let sample
